moving MVC form value route to mvc module
