+++ src/Microsoft.AspNetCore.Mvc.Modules/project.json
@@ -2,6 +2,7 @@
   "version": "1.0.0-*",
   "dependencies": {
     "Microsoft.AspNetCore.Mvc": "1.1.0",
+    "Microsoft.AspNetCore.Modules.Abstractions": "1.0.0-*",
     "Microsoft.AspNetCore.Mvc.Modules.Abstractions": "1.0.0-*",
     "Microsoft.AspNetCore.Routing": "1.1.0",
     "Microsoft.Extensions.DependencyInjection.Abstractions": "1.1.0",
