+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/Controllers/AdminController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Modules;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Localization;
+using Microsoft.AspNetCore.Mvc.Modules.ActionConstraints;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Logging;
 using Orchard.ContentManagement;
@@ -19,7 +20,6 @@ using Orchard.Contents.ViewModels;
 using Orchard.DisplayManagement;
 using Orchard.DisplayManagement.ModelBinding;
 using Orchard.DisplayManagement.Notify;
-using Orchard.Mvc;
 using Orchard.Navigation;
 using Orchard.Settings;
 using YesSql.Core.Services;
@@ -434,7 +434,7 @@ namespace Orchard.Contents.Controllers
         }
 
         [HttpPost, ActionName("Edit")]
-        [Mvc.FormValueRequired("submit.Publish")]
+        [FormValueRequired("submit.Publish")]
         public async Task<IActionResult> EditAndPublishPOST(string contentItemId, string returnUrl)
         {
             var content = await _contentManager.GetAsync(contentItemId, VersionOptions.Latest);
