+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/project.json
@@ -1,17 +1,18 @@
 {
   "version": "2.0.0-*",
   "dependencies": {
+    "Microsoft.AspNetCore.Mvc.Modules": "1.0.0-*",
     "Microsoft.Extensions.Configuration.Json": "1.1.0",
+    "NETStandard.Library": "1.6.1",
     "Orchard.ContentManagement": "2.0.0-*",
     "Orchard.ContentManagement.Display": "2.0.0-*",
-    "Orchard.Settings": "2.0.0-*",
-    "Orchard.Navigation": "2.0.0-*",
-    "Orchard.Indexing.Abstractions": "2.0.0-*",
-    "Orchard.Scripting.Abstractions": "2.0.0-*",
-    "Orchard.Feeds.Abstractions": "2.0.0-*",
-    "Orchard.Recipes.Abstractions": "2.0.0-*",
     "Orchard.ContentTypes.Abstractions": "2.0.0-*",
-    "NETStandard.Library": "1.6.1"
+    "Orchard.Feeds.Abstractions": "2.0.0-*",
+    "Orchard.Indexing.Abstractions": "2.0.0-*",
+    "Orchard.Navigation": "2.0.0-*",
+    "Orchard.Recipes.Abstractions": "2.0.0-*",
+    "Orchard.Scripting.Abstractions": "2.0.0-*",
+    "Orchard.Settings": "2.0.0-*"
   },
   "buildOptions": {
     "define": [ "TRACE" ],
