+++ src/Orchard.Cms.Web/Modules/Orchard.Nancy/project.json
@@ -1,7 +1,6 @@
 ﻿{
   "version": "2.0.0-*",
   "dependencies": {
-    "Microsoft.AspNetCore.Modules": "1.0.0-*",
     "Microsoft.AspNetCore.Nancy.Modules": "1.0.0-*",
     "NETStandard.Library": "1.6.1"
   },
